+++ inmoose/limma/__init__.py
@@ -1,2 +1,3 @@
+from .ebayes import tmixture_matrix, tmixture_vector
 from .fitFDist import fitFDist
 from .squeezeVar import squeezeVar